+++ src/main/java/com/oltpbenchmark/benchmarks/tatp/procedures/GetNewDestination.java
@@ -47,10 +47,10 @@ public class GetNewDestination extends Procedure {
   public void run(Connection conn, long s_id, byte sf_type, byte start_time, byte end_time)
       throws SQLException {
     try (PreparedStatement stmt = this.getPreparedStatement(conn, getNewDestination)) {
-      stmt.setLong(1, s_id);
-      stmt.setByte(2, sf_type);
-      stmt.setByte(3, start_time);
-      stmt.setByte(4, end_time);
+      stmt.setInt(1, (int) s_id);
+      stmt.setInt(2, sf_type);
+      stmt.setInt(3, start_time);
+      stmt.setInt(4, end_time);
       try (ResultSet results = stmt.executeQuery()) {
         assert results != null;
       }
